+++ example/lib/main.dart
@@ -219,6 +219,14 @@ class IntPage extends StatelessWidget {
               );
             },
           ),
+          Subscriber(
+            builder: () {
+              return Text(
+                'sum(computed): $sum',
+                style: Theme.of(context).textTheme.headlineLarge,
+              );
+            },
+          ),
         ],
       ),
     );